# handler/response.go
package handler

import (
	"fmt"
	"net/http"

	"github.com/buzzcoode/gopportunities/schemas"
	"github.com/gin-gonic/gin"
)

func sendError(ctx *gin.Context, code int, msg string) {
	ctx.Header("content-type", "application/json")
	ctx.JSON(code, gin.H{
		"message":   msg,
		"errorCode": code,
	})
}

func sendSucess(ctx *gin.Context, op string, data interface{}) {
	ctx.Header("content-type", "application")
	ctx.JSON(http.StatusOK, gin.H{
		"message": fmt.Sprintf("operation from handler: %s successfull", op),
		"data":    data,
	})
}

type ErrorResponse struct {
	Message   string `json:"message"`
	ErrorCode int    `json:"errorCode"`
}

type CreateOpeningResponse struct {
	Message string                  `json: "message"`
	Data    schemas.OpeningResponse `json: "data"`
}

type DeleteOpeningResponse struct {
	Message string                  `json: "message"`
	Data    schemas.OpeningResponse `json: "data"`
}

type ShowOpeningResponse struct {
	Message string                  `json: "message"`
	Data    schemas.OpeningResponse `json: "data"`
}